Add forceTypeCombinator to generate `type` instead of `interface` declarations
This change adds a `forceTypeCombinator` that takes
`InterfaceCombinator | StrictCombinator | PartialCombinator` as an
argument and ensures that type declarations use `type` intead of
`interface`

In some cases we may prefer to generate type declarations that use
`type` rather than `interface`, but currently `interface` is hard-coded
in cases where it is possible to express the type using `interface`.
This makes perfect sense for the most part, however...

The rules for assignability to a variable that has an
index signature are different when the assignee value is an interface
as opposed to when it is a type.

Currently this is 'as designed':
https://github.com/microsoft/TypeScript/issues/15300#issuecomment-332366024
! Though I don't find the rationale particularly compelling.

To give a concrete example:

This compiles:

```
    type Foo {
      bar: 1;
    }
    const foo: Foo = { bar: 1 };
    let fooContainer: Record<string, number> =  foo
```

But, this does not:
```
    interface Foo {
      bar: 1;
    }
    const foo: Foo = { bar: 1 };
    let fooContainer: Record<string, number> =  foo
```

Whence the last assignment complains.

Signed-off-by: Silas Davis <silas.davis@monaxlabs.io>